refactor(IDX): restructure files
